xiaotiao-server/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
anthropic==0.21.3
python-dotenv==1.0.1
sqlalchemy==2.0.28
pydantic==2.6.4
